+++ BOCCHI.Common/Data/Zones/Graph/Factory/Steps/AddActivitiesStep.cs
@@ -42,9 +42,7 @@ public class AddActivitiesStep : IGraphBuildStep
                 {
                     Id = criticalEncounter.Id,
                     PreferredAethernetId = criticalEncounter.PreferredAethernetId,
-                    CombatRadius = criticalEncounter.CombatRadius ?? 0f,
                     AreaShape = criticalEncounter.AreaShape,
-                    StandRadius = criticalEncounter.StandRadius ?? 0f,
                 }
             });
         }